Move the enum file to correct folder
